requirements.txt
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.3.0.38
pandas==1.1.5
Pillow==9.5.0
